+++ src/SMAPI/ICursorPosition.cs
@@ -6,6 +6,9 @@ namespace StardewModdingAPI
     /// <summary>Represents a cursor position in the different coordinate systems.</summary>
     public interface ICursorPosition : IEquatable<ICursorPosition>
     {
+        /// <summary>The pixel position relative to the top-left corner of the in-game map.</summary>
+        Vector2 AbsolutePixels { get; }
+
         /// <summary>The pixel position relative to the top-left corner of the visible screen.</summary>
         Vector2 ScreenPixels { get; }